+++ src/Controller/TeacherDataController.java
@@ -24,7 +24,6 @@ import Model.GlobalDataModel;
 import Tools.JoAlert;
 import Tools.JoFilechooser;
 import Tools.JoHookEvent;
-import View.HomeView;
 import View.TeacherDataView;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
@@ -162,37 +161,42 @@ public class TeacherDataController implements JoMVC, ActionListener, MouseListen
         TeacherFileService fileService = new TeacherFileService();
         JoAlert alert = new JoAlert();
         if (alert.JoSubmitDelete()) {
-            //===================== Delete History ===================
-            TeacherHistoryModel historyModel = historyService.getTeacherHistoryByTeacherId(TeacherID);
-            historyService.DeleteTeacherHistory(historyModel);
-            //===================== Delete Address =================
-            TeacherAddressModel addressModel = addressService.getTeacherAddressById(TeacherID);
-            addressService.DeleteTeacherAddress(addressModel);
-            //===================== Delete Education ================
-            educationService.getALLByTeacherID(TeacherID).forEach(educationModel -> {
-                educationService.DeleteTeacherEducation(educationModel);
-            });
-            //================== Dlete Experience ================
-            experienceService.getAllExperienceByTeacherID(TeacherID).forEach(experienceModel -> {
-                experienceService.DeleteExperience(experienceModel);
-            });
-            //================= Delete File ===================
-            fileService.getAllTeacherFileByTeacherID(TeacherID).forEach(fileModel -> {
-                fileService.DeleteTeacherFile(fileModel);
-            });
-            //=============== Deletet OutStanding ================
-            outStandingService.getAllByTeacherID(TeacherID).forEach(outstanding -> {
-                outStandingService.DeleteTeacherOutstanding(outstanding);
-            });
-            //=============== Delete Vaccince ================
-            vaccinService.getTeacherVaccinAllByTeacherID(TeacherID).forEach(vaccince -> {
-                vaccinService.DeleteTeacherVaccin(vaccince);
-            });
+            System.out.println(service.getTeacherinFinancial(TeacherID));
+            if (!service.getTeacherinFinancial(TeacherID)) {
+                //===================== Delete History ===================
+                TeacherHistoryModel historyModel = historyService.getTeacherHistoryByTeacherId(TeacherID);
+                historyService.DeleteTeacherHistory(historyModel);
+                //===================== Delete Address =================
+                TeacherAddressModel addressModel = addressService.getTeacherAddressById(TeacherID);
+                addressService.DeleteTeacherAddress(addressModel);
+                //===================== Delete Education ================
+                educationService.getALLByTeacherID(TeacherID).forEach(educationModel -> {
+                    educationService.DeleteTeacherEducation(educationModel);
+                });
+                //================== Dlete Experience ================
+                experienceService.getAllExperienceByTeacherID(TeacherID).forEach(experienceModel -> {
+                    experienceService.DeleteExperience(experienceModel);
+                });
+                //================= Delete File ===================
+                fileService.getAllTeacherFileByTeacherID(TeacherID).forEach(fileModel -> {
+                    fileService.DeleteTeacherFile(fileModel);
+                });
+                //=============== Deletet OutStanding ================
+                outStandingService.getAllByTeacherID(TeacherID).forEach(outstanding -> {
+                    outStandingService.DeleteTeacherOutstanding(outstanding);
+                });
+                //=============== Delete Vaccince ================
+                vaccinService.getTeacherVaccinAllByTeacherID(TeacherID).forEach(vaccince -> {
+                    vaccinService.DeleteTeacherVaccin(vaccince);
+                });
 
-            //============ Delete Teacher Data =================
-            int respon = service.DeleteTeacher(teacherModel);
-            alert.messages("ລົບຂໍ້", "ລົບຂໍ້ມູນອຈານສຳເລັດ", JoAlert.Icons.success);
-            AppTeacher appTeacher = new AppTeacher();
+                //============ Delete Teacher Data =================
+                int respon = service.DeleteTeacher(teacherModel);
+                alert.messages("ລົບຂໍ້", "ລົບຂໍ້ມູນອຈານສຳເລັດ", JoAlert.Icons.success);
+                AppTeacher appTeacher = new AppTeacher();
+            } else {
+                alert.messages("ກວດສອບ", "ຂໍ້ມູນຄູນີ້ຍັງຖືກນຳໃຊ້ໃນຂໍ້ມູນການຈ່າຍຄ່າຮຽນຢູ່", JoAlert.Icons.warning);
+            }
         }
     }
